Added rewrite rule in webpack dev server to proxy /api/auth/complete
Create-react-app provides really nice proxy functionality, but config is
limited to a string: the redirect url. by default, requests with header
'Accept: text/html' will not be proxied. When logging in with Google,
Google redirects to a provided redirect url with a token as a query
parameter. It is dangerous to load a web page with that token because
third-party js can access query parameters. The webserver should
intercept the request and strip out the query parameter. Unfortunately,
the 'Accept' header on the redirect defaults to 'text/html' and so the
request is not proxied, making it impossible to strip the query param.
However, a rewrite rule to bripkens/connect-history-api-fallback solves
this problem. That request now gets proxied.